добавил exceptions

+++ src/main/java/org/example/services/EmployeeService.java
@@ -1,25 +1,20 @@
 package org.example.services;
 
-import org.aspectj.apache.bcel.classfile.Module;
 import org.example.Entity.*;
+import org.example.Exceptions.BusinessException;
+import org.example.Exceptions.EmployeeExistException;
+import org.example.Exceptions.EmployeeNotExistException;
 import org.example.Model.EmployeeModel;
 import org.example.Model.EmployeeResponse;
 import org.example.Model.EmployeeSearchModel;
 import org.example.Repository.*;
-//import org.example.Util.HibernateUtil;
-//import org.hibernate.query.Query;
-import org.example.Util.MapperUtil;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.data.web.SpringDataWebProperties;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -33,27 +28,52 @@ public class EmployeeService {
 
 
     public void saveEmployee(EmployeeModel employeeModel){
+
+        try{
+
         List<EmployeeModel> employeeModelList = new ArrayList<>();
         employeeModelList.add(employeeModel);
         Employee employee = modelInEntity(employeeModelList).get(0);
+        if(employeeRepo.findByNameAndSurnameAndSecondSurnameAndPhoneNumberAndEmail(employee.getName(), employeeModel.getSurname(), employeeModel.getSecondSurname(), employeeModel.getPhoneNumber(), employeeModel.getEmail()) != null){
+            throw new EmployeeExistException("Employee exist");
+        }
         employeeRepo.save(employee);
+        } catch (BusinessException e){
+            throw e;
+        }
 
     }
 
     public EmployeeModel updateEmployee(EmployeeModel employeeModel){
+        try{
         List<EmployeeModel> employeeModelList = new ArrayList<>();
         employeeModelList.add(employeeModel);
         Employee employee = modelInEntity(employeeModelList).get(0);
         employeeRepo.save(employee);
         return employeeModel;
+        } catch (BusinessException e){
+            throw e;
+        }
+
     }
 
     public void delete(Long id){
+        if(!employeeRepo.existsById(id)){
+            throw new EmployeeNotExistException("Employee doesnt exist");
+        }
+        try {
             employeeRepo.deleteById(id);
+        } catch (BusinessException exception){
+            throw exception;
+        }
+
     }
 
 
     public EmployeeResponse findAll(EmployeeSearchModel employeeSearchModel){
+
+        try {
+
         Page<Employee> employees;
 
         if ("".equals(employeeSearchModel.getSurname()) || employeeSearchModel.getSurname() == null){
@@ -80,6 +100,9 @@ public class EmployeeService {
         employeeResponse.setLast(employees.isLast());
 
         return employeeResponse;
+        } catch (BusinessException e){
+            throw e;
+        }
     }
 
     public List<EmployeeModel> entityInModel(List<Employee> employees){
@@ -102,6 +125,7 @@ public class EmployeeService {
     }
 
     public List<Employee> modelInEntity(List<EmployeeModel> employeeModels){
+        try{
         List<Employee> employees = employeeModels.stream()
                 .map(employeeModel -> {
                     Employee employee = new Employee();
@@ -118,9 +142,13 @@ public class EmployeeService {
                 })
                 .collect(Collectors.toList());
         return employees;
+        } catch (BusinessException e){
+            throw e;
+        }
     }
 
     public List<EmployeeModel> pageInModel(Page<Employee> employees){
+        try{
         List<EmployeeModel> employeeDTOs = employees.stream()
                 .map(employee -> {
                     EmployeeModel dto = new EmployeeModel();
@@ -137,6 +165,9 @@ public class EmployeeService {
                 })
                 .collect(Collectors.toList());
         return employeeDTOs;
+        } catch (BusinessException e){
+            throw e;
+        }
     }
 
 }
